Move inheritance information to schema cache
