+++ lib/torque/postgresql.rb
@@ -20,6 +20,7 @@ require 'torque/postgresql/inheritance'
 require 'torque/postgresql/coder'
 require 'torque/postgresql/migration'
 require 'torque/postgresql/relation'
+require 'torque/postgresql/schema_cache'
 require 'torque/postgresql/schema_dumper'
 
 gdep = Gem::Dependency.new('arel', '~> 9.0.0')
